move the forms and templates into their own packages
